+++ environments/dev/common/config/main-local.php
@@ -9,7 +9,7 @@ return [
             'charset' => 'utf8mb4',
         ],
         'mailer' => [
-            'class' => 'yii\swiftmailer\Mailer',
+            'class' => 'yii\symfonymailer\Mailer',
             'viewPath' => '@common/mail',
             // send all mails to a file by default. You have to set
             // 'useFileTransport' to false and configure a transport